% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBBM.R
\name{multiple_GSB}
\alias{multiple_GSB}
\title{A novel biclustering algorithm for mining m6A co-methylation patterns based on beta-binomial distribution and data screening strategy}
\usage{
multiple_GSB(a, b, iteration, burn_in, bic_row_labels, number_bicluster)
}
\arguments{
\item{a}{Ip matrix}

\item{b}{input matrix}

\item{iteration}{Number of initial iterations}

\item{burn_in}{Number of burnings}

\item{bic_row_labels}{Labels for sites in the patterns found, all with initial values of 0}

\item{number_bicluster}{Number of preset patterns}
}
\value{
the patterns hidden in the data
}
\description{
This package enables the discovery of co-methylation patterns for the epigenetic transcription modification m6A methylation. By inputting raw m6A sequencing reads from IP samples and input samples, it automatically uncovers co-methylation patterns hidden within the data.
}
\examples{
rm(list = ls(all = TRUE))
library(EBBM)
data("ip_1")
data("input_1")
number_bicluster <- 6
bic_row_labels <- c(rep(0,dim(ip)[1]))
patterns_found_by_EBBM <- multiple_GSB(a = ip_1,b = input_1,iteration = 1000,burn_in = 500,bic_row_labels = bic_row_labels,number_bicluster = number_bicluster)

}
